dev-dockerise-forecasting-agent: Further clean up; separated unit tests from integration tests
